backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Cache
motor==3.3.2
redis==5.0.1
pymongo==4.6.0

# Object Storage
minio==7.2.0

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
monai==1.3.0
transformers==4.35.2
openai==1.3.7
anthropic==0.7.7

# Medical Imaging
nibabel==5.2.0
pydicom==2.4.4
SimpleITK==2.3.1
opencv-python-headless==4.8.1.78

# NLP & Processing
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# HTTP & Networking
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pillow==10.1.0
